+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-23T02:27:35.252Z",
+    "2026-07-23T11:34:18.983Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {